Revert "push shared json file functionality down to JSONFile"
This reverts commit 806042d1a19cbc6fd72c2149a0e122c0fb107a4f.
